[Phase 02A] Steps 3-7: Security, Repos, Services, Schemas, Rules, API (18 endpoints)
- Security: bcrypt password hashing, JWT (access+refresh with type enforcement), RBAC deps
- 9 repositories: base + User/Product/Order/OrderItem/Logistics/AuditLog/SystemConfig/Idempotency
- 7 services: Auth/Product/Order/Logistics/Audit/Idempotency/SystemConfig
- OrderService: create(aggregate duplicates)/pay(SELECT FOR UPDATE+sorted lock)/cancel(PENDING_PAYMENT only)/ship/deliver
- IdempotencyService: INSERT ON CONFLICT DO NOTHING RETURNING pattern
- State transition validation matrix
- All Pydantic schemas with expected_version fields
- 18 API endpoints across 5 routers (auth 4 + products 4 + orders 7 + logistics 2 + admin 1)
- All monetary fields: NUMERIC(12,2) → Python Decimal
- Ruff: 0 errors. Mypy strict: 0 errors.

+++ backend/app/models/order.py
@@ -2,6 +2,7 @@
 
 import uuid
 from datetime import datetime
+from decimal import Decimal
 
 from sqlalchemy import DateTime, Integer, Numeric, String, Text, func
 from sqlalchemy.dialects.postgresql import UUID
@@ -21,11 +22,11 @@ class Order(Base):
     )
     order_number: Mapped[str] = mapped_column(String(30), unique=True, nullable=False, index=True)
     status: Mapped[str] = mapped_column(String(50), nullable=False, index=True)
-    total_amount: Mapped[float] = mapped_column(Numeric(12, 2), nullable=False)
-    discount_amount: Mapped[float] = mapped_column(Numeric(12, 2), nullable=False, server_default="0")
-    paid_amount: Mapped[float] = mapped_column(Numeric(12, 2), nullable=False, server_default="0")
+    total_amount: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False)
+    discount_amount: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False, server_default="0")
+    paid_amount: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False, server_default="0")
     shipping_address: Mapped[str] = mapped_column(Text, nullable=False)
-    shipping_fee: Mapped[float] = mapped_column(Numeric(12, 2), nullable=False, server_default="0")
+    shipping_fee: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False, server_default="0")
     coupon_code: Mapped[str | None] = mapped_column(String(50), nullable=True)
     paid_at: Mapped[datetime | None] = mapped_column(DateTime(timezone=True), nullable=True)
     shipped_at: Mapped[datetime | None] = mapped_column(DateTime(timezone=True), nullable=True)
